geo_json_rest_api/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-gis
geopandas
psycopg2
django-leaflet
djangorestframework_simplejwt
django-filter